lab_03

+++ src/Lab_02/Lab2_1.java
@@ -3,6 +3,7 @@ package Lab_02;
 import java.util.Scanner;
 
 public class Lab2_1 {
+
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
         System.out.println("height: ");
@@ -14,22 +15,22 @@ public class Lab2_1 {
         BMI = weight / (height * 2);
         System.out.println("Your BMI is: "+BMI+"");
 
-        if (BMI < 18.5){
+        if (BMI < 18.5f){
             System.out.println("Underweight");
             float a;
-            a = (float) (18.5 - BMI);
+            a =  (18.5f - BMI);
             System.out.println("Ban can tang:"+a+"kg");
-        } else if (BMI <= 24.9) {
+        } else if (BMI <= 24.9f) {
             System.out.println("Normal weight");
-        } else if (BMI <= 29.9) {
+        } else if (BMI <= 29.9f) {
             System.out.println("Overweight");
             float b;
-            b = (float) (BMI - 24.9);
+            b = (BMI - 24.9f);
             System.out.println("Ban can phai giam:"+b+"kg");
         } else{
             System.out.println("Obesity");
             float c;
-            c = (float) (BMI - 24.9);
+            c = (BMI - 24.9f);
             System.out.println("Ban can phai giam :"+c+"kg");
         }
     }
